requirements.txt
altair==5.4.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
clip @ git+https://github.com/ultralytics/CLIP.git@17258565f9fc5d170ea4778e5cec03d03f07cbf7
contourpy==1.3.0
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11  # Only if needed for Git integration
GitPython==3.1.43  # Only if needed for Git integration
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.4.1  # If using markdown features
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.10.0.84  # Removed opencv-python
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
scipy==1.14.1
setuptools==74.0.0
sniffio==1.3.1  # If using httpx
streamlit==1.38.0
termcolor==2.4.0
thop==0.1.1.post2209072238
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
ultralytics==8.1.30
urllib3==2.2.2
wcwidth==0.2.13
torch==2.0.0  # Adjust the version as necessary
torchvision==0.19.0  # If you're using torchvision as well